[Bronze II] Title: 숫자, Time: 12 ms, Memory: 3684 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -1,28 +1,30 @@
 #include <string>       // 문자열 처리를 위한 헤더 파일
 #include <iostream>     // 입출력을 위한 헤더 파일
 #include <vector>       // 동적 배열(vector)을 사용하기 위한 헤더 파일
-#include <algorithm>    // 알고리즘 함수를 사용하기 위한 헤더 파일
-#include <stack>        // 스택을 사용하기 위한 헤더 파일
-#include <queue>        // 큐를 사용하기 위한 헤더 파일
-#include <deque>        // 덱(Deque)을 사용하기 위한 헤더 파일
-#include <cctype>       // 문자열의 문자를 검사하기 위한 헤더 파일
-#include <stdlib.h>     // 문자열을 숫자로 변환하는 함수를 사용하기 위한 헤더 파일
+#include <algorithm>    // 알고리즘 함수 (예: min())를 사용하기 위한 헤더 파일
+#include <stack>        // 스택 자료구조를 사용하기 위한 헤더 파일
+#include <queue>        // 큐 및 우선순위 큐(priority_queue)를 사용하기 위한 헤더 파일
+#include <deque>        // 덱(Deque, 양방향 큐)을 사용하기 위한 헤더 파일
+#include <cctype>       // 문자 처리를 위한 헤더 파일
+#include <stdlib.h>     // 문자열을 숫자로 변환하는 함수(예: atoi())를 위한 헤더 파일
 #include <cmath>        // 수학 함수를 사용하기 위한 헤더 파일
-#include <unordered_map> // 해시맵(Unordered Map)을 사용하기 위한 헤더 파일
-#include <map>          // 맵(Map)을 사용하기 위한 헤더 파일
-#include <sstream>      // 문자열을 스트림으로 변환하기 위한 헤더 파일
-#include <cstdlib>      // 일반적인 유틸리티 함수를 사용하기 위한 헤더 파일
-#include <iomanip>      // 입출력 서식 지정을 위한 헤더 파일
-#include <climits> // INT_MAX를 사용하기 위한 헤더 파일
+#include <unordered_map> // 해시맵(unordered_map)을 사용하기 위한 헤더 파일
+#include <map>          // 맵(map, 정렬된 딕셔너리)을 사용하기 위한 헤더 파일
+#include <sstream>      // 문자열 스트림을 사용하기 위한 헤더 파일
+#include <cstdlib>      // 유틸리티 함수(예: rand(), exit())를 위한 헤더 파일
+#include <iomanip>      // 입출력 서식을 지정하기 위한 헤더 파일
+#include <climits>      // INT_MAX(최대 정수값)을 사용하기 위한 헤더 파일
+#include <numeric>      // accumulate(누적 합)를 사용하기 위한 헤더 파일
+#include <limits.h>     // 시스템 한계값들을 정의하는 헤더 파일 (ex: LONG_MAX)
+#include <ctime>
 
-using namespace std;  // 표준 라이브러리(std)의 네임스페이스를 사용함
+using namespace std;
 
-int main()
-{
-	ios::sync_with_stdio(false);	cin.tie(NULL);	cout.tie(NULL);
-    
+int main() {
+    ios::sync_with_stdio(false);cin.tie(NULL);cout.tie(NULL);
+
     unsigned long long a, b; cin >> a >> b;
-    
+
     if (a > b)
     {
         unsigned long long temp = b;
@@ -37,9 +39,9 @@ int main()
     else
     {
         int cnt = 0;
-        vector<int> A;
+        vector<unsigned long long> A;
 
-        for (int i = a + 1; i < b; i++)
+        for (unsigned long long i = a + 1; i < b; i++)
         {
             cnt++;
             A.push_back(i);
@@ -47,12 +49,11 @@ int main()
 
         cout << cnt << "\n";
 
-        for (int i = 0; i < A.size(); i++)
+        for (unsigned long long i = 0; i < A.size(); i++)
         {
             cout << A[i] << " ";
         }
-    }  
-    
+    }
 
     return 0;
 }
